requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pytesseract==0.3.8